+++ History.md
@@ -262,4 +262,11 @@ HomeController 및 테스트 코드를 추가했다.
 
 - h2 데이터베이스 설정 추가
 
+## 3.03 UserRepository 구현 분리
+
+- 데이터베이스 연동을 위해 확장 가능성을 열어야한다.
+- `UserRepository`를 인터페이스로 하고, 기존 사용하던 것은 구현체 MemoryUserRepository로 하였다.
+- 의존 계층에서 역할을 의존하도록 했다.
+- 이제부터 외부에서는 데이터 저장계층의 역할인 userRepository를 의존한다. 구현체의 상태가 바뀌어도 외부의 코드를 변경할 필요가 없다.
+
 ---
